pyproject.toml
[project]
name = "recruiterassist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "ollama>=0.6.1",
    "openpyxl>=3.1.5",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
]
